requirements.txt
streamlit==1.41.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
tenacity==9.0.0
